refactor(repository): 패키지, 테이블명 오타 수정
