+++ src/main/java/com/aliens/backend/chat/domain/repository/ChatRoomRepository.java
@@ -1,4 +1,4 @@
-package com.aliens.backend.chat.domain.ChatRepository;
+package com.aliens.backend.chat.domain.repository;
 
 import com.aliens.backend.chat.domain.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ChatRepository extends JpaRepository<ChatRoom, Long> {
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
     List<ChatRoom> findByMemberId(Long memberId);
 }
