+++ src/main/java/com/aliens/backend/chat/domain/repository/CustomMessageRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.aliens.backend.chat.domain.ChatRepository;
+package com.aliens.backend.chat.domain.repository;
 
 import com.aliens.backend.chat.domain.Message;
 import com.aliens.backend.chat.service.model.ChatMessageSummary;
@@ -17,7 +17,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public class CustomMessageRepositoryImpl implements CustomMessageRepository {
