+++ src/main/java/com/aliens/backend/global/success/ChatSuccessCode.java
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus;
 public enum ChatSuccessCode implements SuccessCode {
     SEND_MESSAGE_SUCCESS(HttpStatus.CREATED, "CH001", "메시지 전송 성공"),
     READ_MESSAGES_SUCCESS(HttpStatus.OK, "CH002", "메시지 읽음 처리 성공"),
+    REPORT_SUCCESS(HttpStatus.CREATED, "CH003", "신고 성공")
     ;
 
     private final HttpStatus httpStatus;
